Always replace User Placeholders with username
MantisBT 2.12.0 changed the way to control the display of users' real
name.

Because of this, user Placeholders were replaced by the realname instead
of the username if the following configuration is used and the current
user's access level is >= $g_show_user_realname_threshold:

    $g_show_realname = ON;
    $g_show_user_realname_threshold = DEVELOPER;

When using such a snippet, real names would be stored in the database,
and thus become visible to any user, even those who are not allowed to
view real names.

To avoid the problem, we now always replace User Placeholders with the
username, regardless of show_realname settings.

Fixes #37

Signed-off-by: Damien Regad <dregad@mantisbt.org>

Original commit message reworded.