Sort Snippets selection list by name
SnippetsPlugin::route_data() method builds an associative array of
snippets (with the snippet's ID as key) which, when converted to JSON,
becomes an Object (i.e. an *unordered* collection of values).

When the SnippetsUI() function (in snippets.js) builds the <select>
list, the Object's contents are processed in ascending order of this
key value, instead of retaining the original sort order, causing the
Snippets to be listed in the wrong order.

This commit removes the PHP array's keys (which are never used anyway),
turning it into a simple array instead of a map. This changes the
resulting JSON into an Array, which is processed in the proper order by
SnippetsUI().

Fixes #34